+++ test/vts/runners/host/test_runner.py
@@ -25,7 +25,10 @@ import os
 import pkgutil
 import signal
 import sys
-import thread
+try:
+    import thread
+except ImportError as e:
+    import _thread as thread
 import threading
 
 from vts.runners.host import base_test
@@ -62,8 +65,8 @@ def main():
     # a test script. The challenge is to handle multiple configs and how to do
     # default config in this case.
     if len(test_classes) != 1:
-        logging.error("Expected 1 test class per file, found %s.",
-                      len(test_classes))
+        logging.error("Expected 1 test class per file, found %s (%s).",
+                      len(test_classes), test_classes)
         sys.exit(1)
     test_result = runTestClass(test_classes[0])
     return test_result
@@ -104,38 +107,51 @@ def runTestClass(test_class):
 
     for config in test_configs:
         if keys.ConfigKeys.KEY_TEST_MAX_TIMEOUT in config:
-            timeout_sec = int(config[keys.ConfigKeys.KEY_TEST_MAX_TIMEOUT]) / 1000.0
+            timeout_sec = int(config[
+                keys.ConfigKeys.KEY_TEST_MAX_TIMEOUT]) / 1000.0
         else:
             timeout_sec = 60 * 60 * 3
             logging.warning("%s unspecified. Set timeout to %s seconds.",
                             keys.ConfigKeys.KEY_TEST_MAX_TIMEOUT, timeout_sec)
-        # The default SIGINT handler sends KeyboardInterrupt to main thread.
-        # On Windows, raising CTRL_C_EVENT, which is received as SIGINT,
-        # has no effect on non-console process. interrupt_main() works but
-        # does not unblock main thread's IO immediately.
-        timeout_func = (raiseSigint if not utils.is_on_windows() else
-                        thread.interrupt_main)
-        sig_timer = threading.Timer(timeout_sec, timeout_func)
+
+        watcher_enabled = threading.Event()
+
+        def watchStdin():
+            while True:
+                line = sys.stdin.readline()
+                if not line:
+                    break
+            watcher_enabled.wait()
+            logging.info("Attempt to interrupt runner thread.")
+            if not utils.is_on_windows():
+                # Default SIGINT handler sends KeyboardInterrupt to main thread
+                # and unblocks it.
+                os.kill(os.getpid(), signal.SIGINT)
+            else:
+                # On Windows, raising CTRL_C_EVENT, which is received as
+                # SIGINT, has no effect on non-console process.
+                # interrupt_main() behaves like SIGINT but does not unblock
+                # main thread immediately.
+                thread.interrupt_main()
+
+        watcher_thread = threading.Thread(target=watchStdin, name="watchStdin")
+        watcher_thread.daemon = True
+        watcher_thread.start()
 
         tr = TestRunner(config, test_identifiers)
         tr.parseTestConfig(config)
         try:
-            sig_timer.start()
+            watcher_enabled.set()
             tr.runTestClass(test_class, None)
         except KeyboardInterrupt as e:
-            logging.exception("Aborted by timeout or ctrl+C: %s", e)
+            logging.exception("Aborted")
         except Exception as e:
             logging.error("Unexpected exception")
             logging.exception(e)
         finally:
-            sig_timer.cancel()
+            watcher_enabled.clear()
             tr.stop()
             return tr.results
-
-
-def raiseSigint():
-    """Raises SIGINT."""
-    os.kill(os.getpid(), signal.SIGINT)
 
 
 class TestRunner(object):
@@ -163,6 +179,7 @@ class TestRunner(object):
                  not.
         test_cls_instances: list of test class instances that were executed
                             or scheduled to be executed.
+        log_severity: string, log severity level for the test logger.
     """
 
     def __init__(self, test_configs, run_list):
@@ -179,7 +196,10 @@ class TestRunner(object):
         l_path = os.path.join(self.test_configs[keys.ConfigKeys.KEY_LOG_PATH],
                               self.testbed_name, start_time)
         self.log_path = os.path.abspath(l_path)
-        logger.setupTestLogger(self.log_path, self.testbed_name)
+        self.log_severity = self.test_configs.get(
+            keys.ConfigKeys.KEY_LOG_SEVERITY, "INFO").upper()
+        logger.setupTestLogger(
+            self.log_path, self.testbed_name, log_severity=self.log_severity)
         self.controller_registry = {}
         self.controller_destructors = {}
         self.run_list = run_list
@@ -310,6 +330,11 @@ class TestRunner(object):
             # in case the controller module modifies the config internally.
             original_config = self.testbed_configs[module_config_name]
             controller_config = copy.deepcopy(original_config)
+            # Add log_severity config to device controller config.
+            if isinstance(controller_config, list):
+                for config in controller_config:
+                    if isinstance(config, dict):
+                        config["log_severity"] = self.log_severity
             logging.info("controller_config: %s", controller_config)
             if "use_vts_agent" not in self.testbed_configs:
                 objects = create(controller_config, start_services)
@@ -321,8 +346,8 @@ class TestRunner(object):
                                "%s, abort!"), module_config_name)
             raise
         if not isinstance(objects, list):
-            raise ControllerError(("Controller module %s did not return a list"
-                                   " of objects, abort.") % module_ref_name)
+            raise signals.ControllerError(("Controller module %s did not"
+                        " return a list of objects, abort.") % module_ref_name)
         self.controller_registry[module_ref_name] = objects
         logging.debug("Found %d objects for controller %s",
                       len(objects), module_config_name)
@@ -338,7 +363,7 @@ class TestRunner(object):
         for name, destroy in self.controller_destructors.items():
             try:
                 logging.debug("Destroying %s.", name)
-                dut = self.controller_destructors[name][0]
+                dut = self.controller_destructors[name]
                 destroy(self.controller_registry[name])
             except:
                 logging.exception("Exception occurred destroying %s.", name)
@@ -387,6 +412,8 @@ class TestRunner(object):
                 cls_result = test_cls_instance.run(test_cases)
             except signals.TestAbortAll as e:
                 raise e
+            finally:
+                self.unregisterControllers()
 
     def run(self):
         """Executes test cases.
